moved state down to the EmailClient level
