+++ frontend/src/components/emails/EmailAttachment.js
@@ -2,6 +2,7 @@ import React from 'react';
 
 export default function EmailAttachment(props) {
   return(
-    <p>EmailAttachment</p>
+    <div className="Email-Attachment">
+    </div>
   )
 }
