+++ frontend/src/components/emails/EmailItem.js
@@ -1,14 +1,14 @@
 import React from 'react';
 import Email from './Email';
 import EmailItemHeader from './EmailItemHeader';
-import EmailAttachments from './EmailAttachments';
+
 
 export default function EmailItem(props) {
+  console.log(props)
   return(
     <div className="EmailItem">
       <EmailItemHeader />
       <Email />
-      <EmailAttachments />
     </div>
   )
 }
