+++ frontend/src/components/emails/EmailListItem.js
@@ -2,15 +2,13 @@ import React from 'react';
 
 export default function EmailListItem(props) {
   return (
-    <div className="EmailListItem" >
+    <div id={props.id} className="EmailListItem" onClick={props.onClick}>
       <h3>{props.from}</h3>
       <h3>{props.subject}</h3>
-      <p>{props.body.substring(0, 30)}</p>
-      <p>{props.timeSent.slice(11).substring(0, 8)}</p>
-      <div className="EmailListItem-IconHolder">
-      <h2>{props.from.substring(0, 1)}</h2>
+      <p>{props.body}</p>
+      <p>{props.timeSent}</p>
+      <h2>{props.from.substring(0,1)}</h2>
       {props.attachments && <i>paperclip</i>}
-      </div>
     </div>
   )
 }
